Refactor and isolate DebugEvent logic
Put it in a separate class, to reduce logic inside ModelActorTransaction
